Module graph_core::link_resolver

Structs

IPFS_TIMEOUT
The timeout for IPFS requests in seconds

LinkResolver

MAX_IPFS_CACHE_FILE_SIZE
The default file size limit for the IPFS cache is 1MiB.

MAX_IPFS_CACHE_SIZE
The default size limit for the IPFS cache is 50 items.

Constants

DEFAULT_MAX_IPFS_MAP_FILE_SIZE
The default file size limit for ipfs.map is 256MiB.

MAX_IPFS_FILE_SIZE_VAR
Environment variable for limiting the ipfs.cat file size limit.

MAX_IPFS_MAP_FILE_SIZE_VAR
Environment variable for limiting the ipfs.map file size limit.

Functions

read_u64_from_env

restrict_file_size

retry_policy

select_fastest_client_with_stat
The IPFS APIs don’t have a quick “do you have the file” function. Instead, we just rely on whether an API times out. That makes sense for IPFS, but not for our application. We want to be able to quickly select from a potential list of clients where hopefully one already has the file, and just get the file from that.