Function graph_core::link_resolver::select_fastest_client_with_stat

async fn select_fastest_client_with_stat(
    clients: Arc<Vec<Arc<IpfsClient>>>, 
    logger: Logger, 
    path: String, 
    timeout: Duration, 
    do_retry: bool
) -> Result<(ObjectStatResponse, Arc<IpfsClient>), Error>

The IPFS APIs don’t have a quick “do you have the file” function. Instead, we just rely on whether an API times out. That makes sense for IPFS, but not for our application. We want to be able to quickly select from a potential list of clients where hopefully one already has the file, and just get the file from that.

The strategy here then is to use the object_stat API as a proxy for “do you have the file”. Whichever client has or gets the file first wins. This API is a good choice, because it doesn’t involve us actually starting to download the file from each client, which would be wasteful of bandwidth and memory in the case multiple clients respond in a timely manner. In addition, we may make good use of the stat returned.